+++ src/blocking_api.rs
@@ -15,7 +15,7 @@
 
 use crate::error::Error;
 use crate::error::Error::{ClientErrorResponse, NotFound, ServerErrorResponse};
-use crate::requests::{EmptyPayload, StreamParams};
+use crate::requests::{Amqp091ShovelParams, EmptyPayload, StreamParams, SHOVEL_COMPONENT};
 use crate::responses::{
     DeprecatedFeatureList, FeatureFlag, FeatureFlagList, FeatureFlagStability, FeatureFlagState,
     GetMessage, OAuthConfiguration, VirtualHostDefinitionSet, WarmStandbyReplicationStatus,
@@ -879,7 +879,10 @@ where
         Ok(response)
     }
 
-    pub fn upsert_runtime_parameter(&self, param: &RuntimeParameterDefinition) -> Result<()> {
+    pub fn upsert_runtime_parameter<'a>(
+        &self,
+        param: &'a RuntimeParameterDefinition<'a>,
+    ) -> Result<()> {
         let _response = self.http_put(
             path!("parameters", param.component, param.vhost, param.name),
             &param,
@@ -1201,6 +1204,33 @@ where
         let response = self.http_get("shovels", None, None)?;
         let response = response.json()?;
         Ok(response)
+    }
+
+    pub fn declare_amqp091_shovel(&self, params: Amqp091ShovelParams<'_>) -> Result<()> {
+        let runtime_param = RuntimeParameterDefinition::from(params);
+
+        let _response = self.http_put(
+            path!(
+                "parameters",
+                SHOVEL_COMPONENT,
+                runtime_param.vhost,
+                runtime_param.name
+            ),
+            &runtime_param,
+            None,
+            None,
+        )?;
+        Ok(())
+    }
+
+    pub fn delete_shovel(&self, vhost: &str, name: &str, idempotently: bool) -> Result<()> {
+        let excludes = if idempotently {
+            Some(StatusCode::NOT_FOUND)
+        } else {
+            None
+        };
+        let _response = self.http_delete(path!("shovels", "vhost", vhost, name), excludes, None)?;
+        Ok(())
     }
 
     //